Fixed typo

+++ PyMEGABASE.py
@@ -378,8 +378,8 @@ class PyMEGABASE:
         chr_gaps_ndx=(gaps[:,0]=='chr'+str(chr))
         for gp in range(len(chr_gaps_ndx)):
             ndx = chr_gaps_ndx[gp]
-            init_loci=np.round(gaps[gp,1]/50000).astype(int)
-            end_loci=np.round(gaps[gp,2]/50000).astype(int)
+            init_loci=np.round(gaps[gp,1].astype(float)/50000).astype(int)
+            end_loci=np.round(gaps[gp,2].astype(float)/50000).astype(int)
             predict_types[init_loci:end_loci]=6
      
         return predict_type
@@ -849,8 +849,8 @@ class PyMEGABASE_extended:
         chr_gaps_ndx=(gaps[:,0]=='chr'+str(chr))
         for gp in range(len(chr_gaps_ndx)):
             ndx = chr_gaps_ndx[gp]
-            init_loci=np.round(gaps[gp,1]/50000).astype(int)
-            end_loci=np.round(gaps[gp,2]/50000).astype(int)
+            init_loci=np.round(gaps[gp,1].astype(float)/50000).astype(int)
+            end_loci=np.round(gaps[gp,2].astype(float)/50000).astype(int)
             predict_types[init_loci:end_loci]=6
